refactor: comp
